scripts/requirements.txt
# Convex HTTP client
convex==0.7.0

# HTTP requests
httpx==0.27.2

# YAML/Markdown frontmatter parsing
python-frontmatter==1.1.0

# Environment variables
python-dotenv==1.0.1

# Date/time utilities
python-dateutil==2.9.0


